pyproject.toml
[project]
name = "marco-qgis-plugin-core"
dynamic = ["version"]
readme = "README.md"
license = "GPL-3.0-only"
license-files = ["LICENSE"]
requires-python = ">=3.12"
authors = [{ name = "Cofactor Oy", email = "dev@cofactor.fi" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]

[tool.uv.workspace]
members = ["components/*"]

[tool.uv]
package = false
default-groups = ["dev", "lint"]

[dependency-groups]
dev = ["prek>=0.3.8"]
docs = [
  "sphinx>=7.0,<9.0",
  "sphinx-rtd-theme>=2.0,<3.0",
  "sphinx-copybutton>=0.5",
  "myst-parser>=3.0,<5.0",
]
lint = [
  "flake8>=6.0.0",
  "flake8-pyproject>=1.2.3",
  "flake8-qgis>=2.0.1,<3.0.0",
  "flake8-spellcheck>=0.28.0",
  "mypy>=1.13.0",
]

[tool.ruff]

lint.ignore = [
  "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
  "D203",    # Conflicts with D211 (no-blank-line-before-class)
  "D213",    # Conflicts with D212 (multi-line-summary-first-line)
  "G002",    # Logging statement uses %
  "PGH003",  # Use specific rule codes when ignoring type issues
  "PLW1641", # eq-without-hash
]

lint.unfixable = [
  "F401", # Unused imports
  "F841", # Unused variables
]

# List of all rules https://docs.astral.sh/ruff/rules/
lint.select = [
  "ANN",  # flake8-annotations
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "C90",  # flake8, mccabe
  "D",    # pydocstyle
  "E",    # flake8, pycodestyle
  "F",    # flake8, Pyflakes
  "FURB", # refurb
  "G",    # flake8-logging-format
  "FBT",  # flake8-boolean-trap
  "I",    # isort
  "ICN",  # flake8-import-conventions
  "INP",  # flake8-no-pep420
  "ISC",  # flake8-implicit-str-concat
  "LOG",  # flake8-logging
  "N",    # pep8-naming
  "PERF", # perflint
  "PIE",  # flake8-pie
  "PGH",  # pygrep-hooks
  "PL",   # pylint
  "PT",   # flake8-pytest-style
  "PTH",  # flake8-use-pathlib
  "RET",  # flake8-return
  "RUF",  # Ruff-specific rules
  "S",    # flake8-bandit
  "SIM",  # flake8-simplify
  "T",    # flake8-print
  "TC",   # flake8-type-checking
  "TID",  # flake8-tidy-imports
  "TRY",  # tryceratops
  "UP",   # pyupgrade
  "W",    # flake8, pycodestyle
]

# Avoiding flagging (and removing) `SC200` from any `# noqa`
# directives, despite Ruff's lack of support for `flake8-spellcheck`.
lint.external = ["SC200", "QGS"]

line-length = 88

target-version = "py312"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
"docs/*" = ["ALL"]
"**/macro_test_utils/*" = ["D"]
"**/test*" = [
  "D",
  "INP001",
  "ANN201",
  "ANN202",
  "E501",
  "PLR2004", # Magic value used in comparison, consider replacing {value} with a constant variable
  "PLR0913", # Too many arguments to function call (len(args) > 5)
  "S101",    # Use of assert detected
  "FBT001",  # Boolean-typed positional argument in function definition
]


[tool.flake8]
max-line-length = 88
spellcheck-targets = "names"
dictionaries = "en_US,python,technical"
select = ["SC", "QGS"]

per-file-ignores = [
  "*/test/*:INP001,SC200,QGS105",
  "*/macro_test_utils/*:INP001,SC200,QGS105",
  "docs/*:SC200",
]

[tool.mypy]
python_version = 3.12
follow_imports = "normal"
show_column_numbers = true
check_untyped_defs = true
namespace_packages = true
ignore_missing_imports = true
exclude = "^test"

[tool.pytest.ini_options]
minversion = "9.0"
addopts = "--import-mode=importlib"
timeout = 1000
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(levelname)-10s (%(filename)s) %(message)s"
log_cli_date_format = "%H:%M:%S"
qt_log_level_fail = "CRITICAL"
qgis_qui_enabled = true

[tool.coverage.paths]
source = ["components/*/src", "**/site-packages"]

[tool.coverage.run]
branch = true
source_pkgs = ["macro_plugin", "qgis_macros"]
omit = []

[tool.coverage.report]
exclude_also = [
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
]
